SudeAI/openai_research_agent/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
openai>=1.50.0
google-generativeai>=0.8.3
duckduckgo-search>=7.2.1
beautifulsoup4>=4.12.3
httpx>=0.28.0
python-dotenv>=1.0.1
jinja2>=3.1.4
pydantic>=2.8.2
